+++ cli/web3cli/MockERC1155.py
